// Dashboard/src/Types/Tables.type.ts
export interface Sms {
    _id: string;
    name: string;
    secretkey: string;
    apikey: string;

}